Add integration test for scrapeRoute
